Added AList and SKD file parsers and a simple calibration script.
alist.AList is a container for alist.AScan objects which store
information from particular scans. When accessing a particular
attribute of an AList object, such as 'source', it returns a list
of that attribute for every scan it contains. More documentation to
come...

parse_skd parses SKD schedule files for scan start times, sources,
and many other parameters. In particular for the phased array it looks
for comments containing the current configuration of the system.

The calibrate script uses the above to parses and is also capable
of parsing SMA Tsys files in order to calibrate the a data set with
Tsys and then plot and print out a formatted file.

I also removed 'padsCN' as it's unrelated.