Added a _ton_ of Exception subclasses and the AbstractList container class.
The exception hierarchy will allow finer control of error handling in user
applications and currently is as follows (more will be added):

Exception
+-- Error
    +-- PivotError
    |   +-- NoPivotError
    |   +-- InconsistentPivotError
    |   +-- PivotNotFoundError
    +-- AbstractScanError
    |   +-- ImproperScanError
    +-- DataOverriddenError

The AbstractList container class will really function more like an ordered
dictionary where the key values are the 'pivots' of its scans. In most cases
this will be a Python datetime object.