AbstractList.__getattr__ now has special meanings for certain attributes.
For example, to get an ordered (by pivot) list of a certain attribute for all
scans in the list just use <instance>.<attribute>, to get this in set form (with
repetitions removed) use <instance>.<attribute>_s, or to get that set converted to
a list use <instance>.<attribute>_sl.

Clearly this can be done by using set() and list() on the first attribute lookup
but this comes in handy when using these attributes in a subclassed list's repr
format. (See 'alist' for an example.)