Added approximate key lookup for AbstactList objects.
If you call an object (use parentheses) with a pivot key as you would
perform a key lookup (with square brackets) what you get is the closest
scan to that key. For example,

>>> pz[:5]
[< 3C273/PZ @ 2010-04-04 05:07:00 >,
 < 3C273/PZ @ 2010-04-04 05:10:00 >,
 < 3C273/PZ @ 2010-04-04 05:13:00 >,
 < 3C273/PZ @ 2010-04-04 05:19:00 >,
 < 3C273/PZ @ 2010-04-04 05:22:00 >]
>>> pz[3]
< 3C273/PZ @ 2010-04-04 05:19:00 >
>>> pz[3].datetime
datetime.datetime(2010, 4, 4, 5, 19)
>>> approx = pz[3].datetime + timedelta(minutes=2)
datetime.datetime(2010, 4, 4, 5, 21)
>>> pz(approx)
< 3C273/PZ @ 2010-04-04 05:22:00 >

Note that pz does not have a scan at time 'approx'. This will be useful
when trying to merge lists that have are at different times (pivots) but
nonetheless can logically be merged (TSys values for example). In the future
I may look into in(ex)traplation of values using this system.