SMATsysList now interpolates Tsys values when merged with another list.
This is kind of hacked and there is probably a better solution, not to
mention it's very slow, but it works. Basically when you merge two lists
by summing them, e.g. pz+sma, where the right-hand list is an SMATsysList
instance, the values for Tsys of all antennas are matched as closely as
possible to the datetimes of the left-hand list.

This is done by creating a merged SMATsysScan object for every datetime in
the left-hand list that aggregates the closest Tsys value for every antenna
to that particular datetime.